Refactor click redirect(and avoid div)
