+++ packages/block-editor/src/components/use-canvas-click-redirect/index.js
@@ -1,0 +1,49 @@
+/**
+ * External dependencies
+ */
+import { overEvery, findLast } from 'lodash';
+
+/**
+ * WordPress dependencies
+ */
+import { useEffect } from '@wordpress/element';
+import { focus, isTextField, placeCaretAtHorizontalEdge } from '@wordpress/dom';
+
+/**
+ * Given an element, returns true if the element is a tabbable text field, or
+ * false otherwise.
+ *
+ * @param {Element} element Element to test.
+ *
+ * @return {boolean} Whether element is a tabbable text field.
+ */
+const isTabbableTextField = overEvery( [
+	isTextField,
+	focus.tabbable.isTabbableIndex,
+] );
+
+export function useCanvasClickRedirect( ref ) {
+	function onClick( event ) {
+		// Only handle clicks directly on the canvas itself, not on content.
+		if ( event.target !== ref.current ) {
+			return;
+		}
+
+		const focusableNodes = focus.focusable.find( ref.current );
+		const target = findLast( focusableNodes, isTabbableTextField );
+
+		if ( ! target ) {
+			return;
+		}
+
+		placeCaretAtHorizontalEdge( target, true );
+	}
+
+	useEffect( () => {
+		ref.current.addEventListener( 'click', onClick );
+
+		return () => {
+			ref.current.removeEventListener( 'click', onClick );
+		};
+	}, [] );
+}
