+++ packages/block-editor/src/components/writing-flow/style.scss
@@ -1,8 +1,3 @@
-.block-editor-writing-flow {
-	display: flex;
-	flex-direction: column;
-}
-
 .block-editor-writing-flow__click-redirect {
 	cursor: text;
 }
