+++ packages/edit-post/src/components/visual-editor/style.scss
@@ -27,17 +27,6 @@
 	}
 }
 
-.editor-styles-wrapper,
-.editor-styles-wrapper > .block-editor-writing-flow__click-redirect {
-	height: 100%;
-}
-
-.edit-post-visual-editor .block-editor-writing-flow__click-redirect {
-	// Allow the page to be scrolled with the last block in the middle.
-	min-height: 40vh;
-	width: 100%;
-}
-
 // Hide the extra space when there are metaboxes.
 .has-metaboxes .edit-post-visual-editor .block-editor-writing-flow__click-redirect {
 	height: 0;
